Fix bug.  Switch should toggle when label is clicked.
In the off state, if you click on the label the switch does not toggle.
 This is because the offset is non-zero, which makes the difference
non-zero, which makes switch behave as if it’s been dragged to the off
state.

The fix is to only calculate the state based on the difference if the
switch was dragged.  And it’s only dragged if the mouse has moved at
some point after the initial dragStart.

I took a look at the original bootstrap switch and it’s how it behaves
too.