Do not fill 100% of parent height
When using the switch, I get a massive height.  It’s the same as the
issue reported here:
https://github.com/nostalgiaz/bootstrap-switch/issues/360

As I understand it, height specifies the height of an element relative
to its parent container, so 100% is the same height as its parent.
http://www.w3.org/TR/CSS21/visudet.html#propdef-height

However, I think it makes sense for a switch to be defined by its
internal height (as dictated by its size / content).