fixed bug #6 with switch not toggling with state prop update
Previously when the state prop changed due to an update in the parent component, the switch would not toggle to the correct position.  Now, with the use of componentWillReceiveProps, we pass the state prop to the value() method so it can be updated.  This will probably work with other, currently immutable props.